review fixes #6
