+++ src/main/java/ru/hh/resumebuilderbot/cv/builder/CVBuilder.java
@@ -1,0 +1,7 @@
+package ru.hh.resumebuilderbot.cv.builder;
+
+import ru.hh.resumebuilderbot.ChatId;
+
+public interface CVBuilder {
+    String build(ChatId chatId);
+}
