+++ src/main/java/ru/hh/resumebuilderbot/cv/builder/CVFormats.java
@@ -1,0 +1,12 @@
+package ru.hh.resumebuilderbot.cv.builder;
+
+public enum CVFormats {
+    PLAIN_TEXT {
+        @Override
+        public CVBuilder getBuilder() {
+            return new PlainTextCVBuilder();
+        }
+    };
+
+    public abstract CVBuilder getBuilder();
+}
