+++ src/main/java/ru/hh/resumebuilderbot/question/generator/ShowAllQuestionGenerator.java
@@ -1,14 +1,14 @@
 package ru.hh.resumebuilderbot.question.generator;
 
-import ru.hh.resumebuilderbot.CVFormats;
 import ru.hh.resumebuilderbot.ChatId;
 import ru.hh.resumebuilderbot.Question;
+import ru.hh.resumebuilderbot.cv.builder.CVFormats;
 
 public class ShowAllQuestionGenerator implements QuestionGenerator {
 
 
     @Override
     public Question generateNext(ChatId chatId) {
-        return new Question(chatId, CVFormats.PLAIN_TEXT.build(chatId));
+        return new Question(chatId, CVFormats.PLAIN_TEXT.getBuilder().build(chatId));
     }
 }
